% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cache}
\usage{
clear_cache(dataset = NULL)
}
\arguments{
\item{dataset}{Optional. Specific dataset to clear, or NULL for all}
}
\description{
Clear cache
}
