% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_dataset.R
\name{print_citation}
\alias{print_citation}
\title{Print citation information}
\usage{
print_citation(name, year = NULL, server = NULL)
}
\arguments{
\item{name}{Character. Dataset identifier (DOI, alias, or metadata ID)}

\item{year}{Numeric. Year (if applicable)}

\item{server}{Character. Dataverse server URL (optional)}
}
\description{
Print citation information
}
