% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_dataset.R
\name{print_citation}
\alias{print_citation}
\title{Print citation information}
\usage{
print_citation(name, year = NULL)
}
\arguments{
\item{name}{Character. Dataset name}

\item{year}{Numeric. Year (if applicable)}
}
\description{
Print citation information
}
